add some room for future flexiblity with integration tests
