+++ package.json
@@ -53,6 +53,7 @@
     "eslint-plugin-unused-imports": "~2.0.0",
     "eventemitter3": "^4.0.7",
     "fastify": "~4.14.1",
+    "flik": "0.4.2",
     "husky": "8.0.3",
     "is-visible": "^2.2.0",
     "karma": "^6.4.0",
@@ -63,7 +64,7 @@
     "karma-spec-reporter": "^0.0.34",
     "karma-typescript": "~5.5.3",
     "lint-staged": "^13.2.0",
-    "melodist": "0.8.0",
+    "melodist": "0.8.1",
     "mocha": "^10.0.0",
     "node-fetch": "2.6.9",
     "npm-run-all": "^4.1.3",
