+++ test/integration/login/fixtures.json
@@ -1,0 +1,56 @@
+[
+  {
+    "name": "DigitalOcean",
+    "url": "https://cloud.digitalocean.com/login",
+    "expectedFields": {
+      "username": "input#email",
+      "password": "input#password"
+    }
+  },
+  {
+    "name": "GitHub",
+    "url": "https://github.com/login",
+    "expectedFields": {
+      "username": "input#login_field",
+      "password": "input#password"
+    }
+  },
+  {
+    "name": "DynDNS",
+    "url": "https://account.dyn.com/entrance/",
+    "expectedFields": {
+      "username": "input[id^=login][name=username]",
+      "password": "input[id^=login][type=password]"
+    }
+  },
+  {
+    "name": "Reddit",
+    "url": "https://www.reddit.com/login",
+    "expectedFields": {
+      "username": "#loginUsername",
+      "password": "#loginPassword"
+    }
+  },
+  {
+    "name": "Amazon",
+    "url": "https://www.amazon.com/ap/signin?_encoding=UTF8&ignoreAuthState=1&openid.assoc_handle=usflex&openid.claimed_id=http%3A%2F%2Fspecs.openid.net%2Fauth%2F2.0%2Fidentifier_select&openid.identity=http%3A%2F%2Fspecs.openid.net%2Fauth%2F2.0%2Fidentifier_select&openid.mode=checkid_setup&openid.ns=http%3A%2F%2Fspecs.openid.net%2Fauth%2F2.0&openid.ns.pape=http%3A%2F%2Fspecs.openid.net%2Fextensions%2Fpape%2F1.0&openid.pape.max_auth_age=0&openid.return_to=https%3A%2F%2Fwww.amazon.com%2F%3Fref_%3Dnav_custrec_signin&switch_account=",
+    "expectedFields": {
+      "username": "form[name=signIn] input#ap_email"
+    }
+  },
+  {
+    "name": "New Relic",
+    "url": "https://rpm.newrelic.com/",
+    "expectedFields": {
+      "username": "#login_email"
+    }
+  },
+  {
+    "name": "giffgaff",
+    "url": "https://www.giffgaff.com/auth/login",
+    "expectedFields": {
+      "username": "#memberName",
+      "password": "#password"
+    }
+  }
+]
